refactor(API): Filters: basculer dans un fichier utils.py dédié
